# os-scheduler-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这是一个操作系统调度实验的后端API服务",
        "title": "操作系统调度器 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/process": {
            "post": {
                "description": "添加一个新的进程到系统中",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "添加新进程",
                "parameters": [
                    {
                        "description": "进程信息",
                        "name": "process",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PCB"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PCB"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            }
        },
        "/processor-status": {
            "get": {
                "description": "获取所有处理机的当前运行状态，包括每个处理机上正在运行的进程信息",
                "produces": [
                    "application/json"
                ],
                "summary": "获取处理机状态",
                "responses": {
                    "200": {
                        "description": "获取处理机状态成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/main.ProcessorStatusResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/reset": {
            "post": {
                "description": "强制重启整个系统，清空所有进程和内存。这将终止所有正在运行的进程，释放所有内存分配，并将系统恢复到初始状态。系统参数（如处理机数量、最大进程数）将保持不变。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "重置系统",
                "responses": {
                    "200": {
                        "description": "系统重置成功",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "500": {
                        "description": "系统重置失败",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            }
        },
        "/resume/{pid}": {
            "post": {
                "description": "恢复已挂起的进程，使其重新参与调度",
                "produces": [
                    "application/json"
                ],
                "summary": "恢复进程",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "进程ID",
                        "name": "pid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "进程恢复成功",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "400": {
                        "description": "进程恢复失败",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            }
        },
        "/schedule": {
            "post": {
                "description": "执行一次进程调度，更新进程状态和处理机分配",
                "produces": [
                    "application/json"
                ],
                "summary": "执行调度",
                "responses": {
                    "200": {
                        "description": "调度执行成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.ProcessQueue"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/status": {
            "get": {
                "description": "获取当前系统的状态信息，包括进程队列和内存管理状态",
                "produces": [
                    "application/json"
                ],
                "summary": "获取系统状态",
                "responses": {
                    "200": {
                        "description": "获取状态成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/main.StatusResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/suspend/{pid}": {
            "post": {
                "description": "将指定进程挂起，暂停其执行",
                "produces": [
                    "application/json"
                ],
                "summary": "挂起进程",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "进程ID",
                        "name": "pid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "进程挂起成功",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "400": {
                        "description": "进程挂起失败",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.ProcessorStatusResponse": {
            "type": "object",
            "properties": {
                "processors": {
                    "description": "每个处理机当前运行的进程，如果没有则为nil",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PCB"
                    }
                }
            }
        },
        "main.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码，0表示成功，非0表示错误",
                    "type": "integer"
                },
                "data": {
                    "description": "数据，可以为空"
                },
                "message": {
                    "description": "提示信息",
                    "type": "string"
                }
            }
        },
        "main.StatusResponse": {
            "type": "object",
            "properties": {
                "memory": {
                    "$ref": "#/definitions/models.MemoryManager"
                },
                "queue": {
                    "$ref": "#/definitions/models.ProcessQueue"
                }
            }
        },
        "models.MemoryBlock": {
            "type": "object",
            "properties": {
                "isUsed": {
                    "type": "boolean"
                },
                "length": {
                    "type": "integer"
                },
                "start": {
                    "type": "integer"
                }
            }
        },
        "models.MemoryManager": {
            "type": "object",
            "properties": {
                "blocks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MemoryBlock"
                    }
                },
                "osSize": {
                    "type": "integer"
                },
                "totalSize": {
                    "type": "integer"
                }
            }
        },
        "models.PCB": {
            "type": "object",
            "properties": {
                "memorySize": {
                    "type": "integer"
                },
                "memoryStart": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "pid": {
                    "type": "integer"
                },
                "predecessors": {
                    "description": "前驱进程PID列表",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "priority": {
                    "type": "integer"
                },
                "processorId": {
                    "description": "-1表示未分配处理机",
                    "type": "integer"
                },
                "requiredTime": {
                    "description": "剩余运行时间",
                    "type": "integer"
                },
                "state": {
                    "$ref": "#/definitions/models.ProcessState"
                },
                "successors": {
                    "description": "后继进程PID列表",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "totalTime": {
                    "description": "总运行时间",
                    "type": "integer"
                }
            }
        },
        "models.ProcessQueue": {
            "type": "object",
            "properties": {
                "backup": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PCB"
                    }
                },
                "ready": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PCB"
                    }
                },
                "running": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PCB"
                    }
                },
                "suspended": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PCB"
                    }
                },
                "waiting": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PCB"
                    }
                }
            }
        },
        "models.ProcessState": {
            "type": "string",
            "enum": [
                "ready",
                "running",
                "waiting",
                "finished",
                "suspended"
            ],
            "x-enum-varnames": [
                "Ready",
                "Running",
                "Waiting",
                "Finished",
                "Suspended"
            ]
        }
    }
}